backend-flask/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
requests==2.32.3

# Local LLM Support
ollama==0.4.4
langchain==0.3.15
langchain-community==0.3.15

# Vector Embeddings & RAG
sentence-transformers==3.3.1
chromadb==0.5.23

# Data Processing
pandas==2.2.3

# Database & Auth
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
email-validator==2.1.0
werkzeug==3.1.3
